changed where hovermode lives
