+++ web/src/features/fireWeather/components/graphs/TempRHGraph.tsx
@@ -1,9 +1,6 @@
 /* eslint-disable @typescript-eslint/ban-ts-comment */
 import React, { useEffect, useState } from 'react'
 import Plot from 'react-plotly.js'
-import FormControl from '@material-ui/core/FormControl'
-import Select from '@material-ui/core/Select'
-import MenuItem from '@material-ui/core/MenuItem'
 import { ObservedValue } from 'api/observationAPI'
 import { ModelValue, ModelSummary } from 'api/modelAPI'
 import { NoonForecastValue, ForecastSummary } from 'api/forecastAPI'
@@ -53,6 +50,7 @@ interface Props {
   rdpsSummaries: ModelSummary[]
   gdpsModels: ModelValue[]
   gdpsSummaries: ModelSummary[]
+  hoverMode: 'closest' | 'x' | 'x unified'
 }
 
 const TempRHGraph = (props: Props) => {
@@ -70,7 +68,8 @@ const TempRHGraph = (props: Props) => {
     gdpsModels,
     gdpsSummaries,
     rdpsModels,
-    rdpsSummaries
+    rdpsSummaries,
+    hoverMode
   } = props
 
   const obsGraphProperties: TempRHGraphProperties = {
@@ -163,18 +162,6 @@ const TempRHGraph = (props: Props) => {
 
   // Update plotly revision to trigger re-drawing of the plot
   const setRevision = useState(0)[1]
-  const [hoverMode, setHoverMode] = useState<'closest' | 'x' | 'x unified'>('closest')
-
-  const handleHoverModeChange = (
-    event: React.ChangeEvent<{ name?: string; value: unknown }>
-  ) => {
-    switch (event.target.value) {
-      case 'closest':
-      case 'x':
-      case 'x unified':
-        setHoverMode(event.target.value)
-    }
-  }
 
   useEffect(() => {
     setRevision(revision => revision + 1)
@@ -182,20 +169,6 @@ const TempRHGraph = (props: Props) => {
 
   return (
     <div id="temp-rh-graph" data-testid="temp-rh-graph">
-      <div style={{ textAlign: 'right', paddingRight: '30px', lineHeight: '40px' }}>
-        Hover mode:{' '}
-        <FormControl size="small">
-          <Select
-            variant="outlined"
-            value={hoverMode as string}
-            onChange={handleHoverModeChange}
-          >
-            <MenuItem value={'closest'}>closest</MenuItem>
-            <MenuItem value={'x'}>x</MenuItem>
-            <MenuItem value={'x unified'}>x unified</MenuItem>
-          </Select>
-        </FormControl>
-      </div>
       <Plot
         style={{ width: '100%', height: '100%' }}
         useResizeHandler={true}
